net/qt5-enginio: deprecate, schedule for removal on 2016-08-04
The Qt Enginio port was introduced with the Qt 5.5.1 version to
ports, however, upstream development is discontinued and the upstream
code is marked as deprecated in the Qt 5.6 version and removed in
Qt 5.7. As the port is currently not used by any other port, we
would like to schedule the removal of our port already for our
upgrade to Qt 5.6. If any interested parties would still like to
use the port, you are given time until August 4, 2016 to notice
kde@FreeBSD.org for a possible solution. Otherwise the port will
get removed at latest by importing Qt 5.6.x to ports.

PR:		210015
Submitted by:	Ralf Nolden <nolden@kde.org> (kde)