New port: sysutils/bstack
bstack - gstack for FreeBSD

This is a port of gstack, a script build around gdb on RedHat and other
Linux distros, to FreeBSD.

bstack take a PID as only parameter and print out the back trace of that
running process. If the process is multithread, the backtrace of all
threads is printed.

It works with the last version of GDB

WWW: https://github.com/pizzamig/bstack

PR:		209398
Submitted by:	luca.pizzamiglio@gmail.com
Approved by:	jadawin (mentor)
Differential Revision:	https://reviews.freebsd.org/D6899