ports-mgmt/synth: Upgrade version 1.40 => 1.41
This release adds detection and handling of circular dependencies.
Two users that tried replacing OpenSSL with LibreSSL found out that
ftp/curl depends on itself when its built with GNUTLS.  That's because
Synth goes into an infinity dependency loop, runs out of stack space,
and crashes.

Now Synth can detect a circular dependency and abort gracefully and
informatively, e.g.

> synth status ftp/curl
ftp/curl scan aborted because a circular dependency on ftp/curl was detected.
... backtrace devel/cmake
... backtrace emulators/tpm-emulator
... backtrace security/trousers
... backtrace security/gnutls
... backtrace ftp/curl

While here, build synth with stack-checking on so out-of-stack situations
result in an exception and not an illegal instruction crash.