MFH: r418191
Take maintainership

Details:
mkvtoolnix shows spurious build issues due to a gcc-internal segfault
on the build cluster on 9.x. However, this is not perfectly
reproducible and on other hardware setups it builds reliably, so we
don't want to mark it broken on 9.x.
Taking maintainership so pkg-fallout won't continue spam the
multimedia@ mailing list.

Approved by:	ports-secteam (junovitch)