Add missing dependencies reported by stage-qa:         USE_XORG+=      sm xinerama
        devel/dbus (when GNOME option is enabled)

        multimedia/gstreamer (when MMEDIA option is enabled)

        graphics/poppler (-devel only, when PDFIMPORT option is enabled)

Add USES=ssl since this OpenOffice does use OpenSSL.  The base and ports
versions of OpenSSL are both known to work.  It is unknown if LibreSSL
works because the dependency ftp/curl does not currently build with
LibreSSL.

Replace one remaining path to a .jar file with ${JAVALIBDIR} in
CONFIGURE_ARGS.

Replace an absolute symlink with a relative one.

Re-align \ line continuation characters in *_DEPENDS after removal of
${PORTSDIR} from dependencies, and make a few other whitespace cleanups.

MFH:		2016Q3