Update to 7.5. While here, completely rewrite the Makefile. Use sed instead of hard patches to minimize maintainership, and drastically simplify the installation process. Add some OPTIONS to handle dependencies on country lookups.
Previously, the path to the icons was being manually changed from
".../icon" to ".../icons". With this update, the port now installs
into the default location (.../icon). The awstatsicons alias will
need to be manually updated for existing installations; this is
explained in UPDATING.

Also, I am taking this port from the previous maintainer. It has
been over two years since he responded to a PR, and this is the
third in a row that has had no response. Under section 5.5 of the
PHB, I am considering this port to be abandoned. I'd like to thank
Alex for all the time that he did spend maintaining this port.

PR:		210821
Approved by:	maintainer timeout (2 weeks)