Remove the SSL option and make it mandatory.
I received a report that the build is broken on 9.x with SSL disabled.
I think the right solution here is just to remove the SSL option entirely.
I really can't think of a use-case where IMAP/POP/LMTP without TLS
capability is a good idea. Plus, NetBSD and OpenBSD both have SSL
mandatory.

PORTREVISION bump is necessary here, because it was possible to have
no SSL on 10.x with a successful build.

Discussed with:	Larry Rosenman, maintainer of dovecot2-pigeonhole