devel/adacurses: Remove hardcoded dynamic linking, plus ...
The AdaCurses source contains "pragma Linking_Options" which hardcodes
linking flags like "-lncurses" and "-lmenu".  This makes it very hard
to statically link libncurses because the pragma flags come at the
end causing the linker to include these libraries.  Fix it by removing
the pragmas and require the user to specify the flags themselves.

Related: the provided adacurses.gpr was also hardcoded for dynamic linking
but it only included the base library (excluding menu, form, and panel).
Rework this gpr file to continue to link dynamically by default, and
continue to exclude menu, for, and panel by default, but add environment
variables that easily allow static linking and adding of menu/form/panel
individually as required.