print/lilypond-devel: mark broken, deprecate for removal after 31 August
This unstable branch of lilypond is already several releases behind.  A
devel port should be required to have an active maintainer, yet this one
does not.  Now that it's broken by a very astonishingly bad pkgconf
detection code, set it for removal in September.  That will give someone
a chance to volunteer to maintain it.