archivers/file-roller: revisit r406930 (remove unzip)
Back in January, I was working under the idea that the ports unzip was
desired, but now it's clear that file-roller was just one of many ports
that had a redundant requirement on unzip.  Revert my January commit
and remove the requirement on unzip completely (it's provided by all
supported releases of FreeBSD).

Approved by:	infrastructure blanket (removal of redundant deps)