deskutils/freemind: Remove redundant dependency on unzip
A dependency on unzip will never be registered since unzip is available
on all supported platforms (since FreeBSD 8.0).
Note that it's pulled in by a non-default option.

In this particular case, USES=zip:infozip is set, so unzip is pulled in
anyway through this mechanism (so explicit callout is still redundant)

Approved by:	infrastructure blanket (removal of redundant deps)