graphics/qcomicbook: Remove redundant dependency on unzip
A dependency on unzip will never be registered since unzip is available
on all supported platforms (since FreeBSD 8.0).
Note that it's pulled in by a non-default option.

In this port, unzip was a non-default (and non-effective) option.
Removing the option requires a revbump.

Approved by:	infrastructure blanket (removal of redundant deps)