net/ipdecap: Fix LIB_DEPENDS on ports libpcap
It is clear from the build dependency on the ports libpcap header that
ipdecap intends to have a dependency on ports libpcap.  However, it is
ambiguously specified and the base libpcap satisfies the requirement, thus
the dependency is never registered.  Use the full library name (so.1 which
does not conflict with either FreeBSD or DF base) to remove the ambiguity.

Approved by:	just-fix-it