New port: x11-fonts/hack-font.
Hack is monospaced font designed to be a workhorse typeface for code. It has
deep roots in the libre, open source typeface community and expands upon the
contributions of the Bitstream Vera & DejaVu projects. The project is in
active development. We welcome your input and contributions.

WWW: https://sourcefoundry.org/hack/