WIP: external toolchain cross buildable: binutils
Add a cross buildable binutils package.
The new category is not linked to the regular ports tree to avoid make install,
poudriere and others to catch it automagically

instead of ending with a very complex file removal in the stage, prefer to use
specific plist per arch.

For now only sparc64 tested and added. This version of binutils is stipped down
only the components that are not supported by elftoolchain