Add x11/xfce4-goodies: Meta-port for ports from the Xfce4 Goodies Project
This meta-port gathers together each of the ports from the Xfce4 Goodies
Project and allows them to be easily installed to create a functional
desktop environment.

Default options include those which do not bring in too many additional
dependencies, and are generally recommended for use on a modern Xfce
desktop.

It also includes ports which are not officially part of the Xfce4
Goodies Project (can not be found at http://goodies.xfce.org/),
but are generally recommended for use with Xfce4 on FreeBSD.

This goal of this meta-port is to help users get started with Xfce by:
1. Easily installing a sensible list of programs with a single command
2. Find new programs they may be interested in, by reviewing the list
   of disabled by default options in the port
3. Show people which programs are not recommended for use by the FreeBSD
   Xfce team (such as obsolete applications)

Reviewed by:	olivierd (with Xfce hat)
Approved by:	sbruno (with Xfce hat), mat (mentor)
Differential Revision:	https://reviews.freebsd.org/D6981