net-mgmt/ndpmon: USES+=SSL and port options overhaul
The py4suite option probably didn't work even before the py4suite
dependency was removed from the tree 2 years ago.

The MACRESOLV option fetches during the build which is illegal and
maybe is hitting a bad URL because it freezes as well.

The py4suite has been removed completed and the MACRESOL option has
been marked broken.  This is after I rearranged the port to make them
even work as intended.  I cleaned up the port a lot after documenting
the SSL requirement with USES=ssl.

Approved by:	just-fit-it and SSL blankets