devel/hadoop: Unbreak build (nominal case)
The ssl.mk helper can augment the MAKE_ARGS variable with a OPENSSL_CFLAGS
definition which just happens to break this port.  Add a filter to remove
any such definitions if they occur.  The issue was introduced when
USES+=ssl was added to the makefile to document the SSL requirement.