misc/fortune-mod-culmea-culmilor: Use fortune_strfile instead of strfile(8)
This removes the dependency on the games distribution.
While here, unsuppress build and installation commands for all ports
that pull in generic makefile fragment.

Unfortunately the tecnik93.com server is down.  I don't know if its
transient or if its been down a long time, but mark the port BROKEN
(unfetchable) as a result.

Approved by:	general blanket