New port: devel/libgrading Library for grading C/C++ assignments
This port provides libgrading, a library for evaluating C- and C++-language
assignments in programming classes. It could also be used as a unit testing
framework, but it's designed to treat the code under test with some
suspicion (using Capsicum sandboxing by default). It also allows different
tests to be assigned different weights towards the overall grade/score,
which is not typically a feature of unit testing frameworks.

Approved by:	mat
Differential Revision:	https://reviews.freebsd.org/D7956