databases/mariadb101: Fix stage-QA for client
THere were 5 new programs built and installed by both the client and
the server.  This is not a unique case.  Previously they were allocated
to the client and removed by the post-install target of the server. In
this case, they've already been allocated to the server, so we need a
new post-install target for the client to remove them from the stage
directory so stage-QA checks pass.

While here, remove the now-unnecessary DragonFly configure argument
and change the backtick in a comment to an apostrophe to avoid the
makefile being interpreted as a shell script by some editors.  Also
remove redundant forward slashes on the server post-install target.

Approved by:	just-fix-it