devel/git: Fix build with PERL option disabled
configure.ac defines a macro (GIT_ARG_SET_PATH), which takes an optional
'allow-without' parameter, which if not specified (in configure.ac),
causes --without-foo/--with-foo=no not to be checked/allowed.

allow-without is set for python, but not for perl in configure.ac:

 # Define PYTHON_PATH to provide path to Python.
 # GIT_ARG_SET_PATH(python, allow-without)

 # GIT_ARG_SET_PATH(perl)

The non-presence of the optional second macro argument results in the
following error when --without-perl is specified by
PERL_CONFIGURE_WITH=perl=${PERL}

  configure:4065: error: You cannot use git without perl

This error is a regression after r427505 [1], which switched from
PERL_MAKE_ENV to PERL_CONFIGURE_WITH (and its associated --without-*).

This change adds:

 * allow-without to configure.ac allowing --without-perl to work.
 * USES=autoreconf to regenerate configure from configure.ac.

[1] https://svnweb.freebsd.org/changeset/ports/427505

PR:			215038
Approved by:		portmgr (blanket)
Differential Revision:	D8709