Update to patchlevel 130.
Patch 129 was a fix for parallel make. It builds fine for me on
FreeBSD with -j4, and on macOS with -j8, but that's the extent
of what I can test on my own. I'm removing MAKE_JOBS_UNSAFE with
this commit, but if one of you with your crazy 256-core machines
encounters build failures then please let me know!