xen-tools: fix path of xenstore device
When xenstore is not running on the same domain as the toolstack the xenstore
device is used in order to access it (like when running some tools from an
unprivileged domain). Fix the path to point to the FreeBSD path
(/dev/xen/xenstore). This is a backport of upstream commit 7ff99b.

Sponsored by:	Citrix Systems R&D
Approved by:	bapt
MFH:		2016Q4
--This line, and hose below, will be ignored--

M    sysutils/xen-tools/Makefile
AM   sysutils/xen-tools/files/0001-libs-xenstore-set-correct-FreeBSD-device.patch