Fix py-pykde4 against newest clang
PyList_SET_ITEM() is supposed to be void, but as a macro it's an
expression and has a pointer value. Clang 4.0 trips on the bogus
pointer comparison; given the comparison, it should be PyList_SetItem()
which returns -1 on failure.

As the produced code may be changed by this, bump PORTREVISION.

PR:		216253
Submitted by:	Adriaan de Groot <groot@kde.org>
Reported by:	jbeich
Reviewed by:	rakuco, tcberner
Approved by:	rakuco (mentor)
Differential Revision:	https://reviews.freebsd.org/D9253