devel/cargo: Fix bootstrap checksum + use ${OPENSSLBASE}
Because devel/cargo and lang/rust-nightly use a different DISTFILE with
the same name, the wrong checksum and size was selected in devel/cargo.
To fix this, keep the parent directory (a date) in the DISTFILE name.

Also, depending on the value of DEFAULT_VERSION=ssl= and whether OpenSSL
from ports is installed, the openssl-sys crate could fail to build.
Passing OpenSSL location to make was way too complicated anyway. So
simplify it and use ${OPENSSLBASE}. This was reported in PR 216385.

Bump PORTREVISION to 1.

PR:		216143, 216385
Reported by:	gja822@narod.ru (216385, SSL build failure)
Approved by:	antoine (mentor), jbeich
Differential Revision:	https://reviews.freebsd.org/D9295