ports-mgmt/synth: Upgrade version 1.66 => 1.67
This version blocks TTY signals to and from background processes, allowing
lang/sbcl to build as well as avoiding disruption of an active ncurses
display by "you've got mail" message.

It also includes a minor duration time formatting fix on build logs and
the many-thousands-open-file-descriptor problem fix on purge-distfiles
command provided by 1.66_1 between releases.

There is also a change in the build order algorithm.  Previously the
categories were ordered alphabetically, but rather scanned in the order
that appears by walking the directory.  Now when the "nothing depends on
these ports" ports block starts building, it's in A-Z order by category/
portname whereas before ony the portname was ordered A-Z (the categories
were blocks in random order).  There is no real benefit to the reordering;
it's mainly to satisfy those bothered by disorder.