Fix i386 SSE2 crashes (use clang), and clean up a bit.
GCC 5.4 also does not get the stack properly aligned in i386 for SSE2
and causes SIGBUS on startup when trying to access an 8-byte aligned
double[8] array with SIMD instructions.[1] One workaround would be to use
-mstackrealign, the other is to use clang - we don't get OpenMP on i386
anyhow.

Move DISTVERSIONSUFFIX up to appease portlint.

Drop -fexpensive-optimizations, clang does not support it, and gcc 5.4
includes it in -O2 and -O3.

Note that the upstream discourages the use of 32-bit architectures for
rawtherapee.

Reported by:	pkg-fallout@ [1]