On amd64, clang++ 4.0 uses the MOVAPS SSE instruction to initialize sufficiently large objects returned by the "new" operator.  This requires that the object have 16-byte alignment.  The FreeBSD malloc() implementation does the correct thing here, but OpenOffice has a couple of internal memory allocator implementations that only align to 8-byte boundaries at most.  In addition OpenOffice overrides the new operator to interpose a couple of layers of wrappers.  If the --enable-debug option is passed to configure, the wrapper adds 8 to the size passed to the allocator and adds an 8 byte offset to the pointer returned by the allocator to make room for a signature that it adds to the beginning of the memory block (the signature is validated and the inverse transformation is done when the memory is freed).  This breaks the proper alignment done by the mamory allocator.  Fix these problems by adding an EXTRA_PATCH that teaches the internal OpenOffice memory allocators to do 16-byte alignment and to use a 16-byte offset in the "new" wrapper, and apply this patch on amd64 if clang 4.0 is the system compiler. Pass the --with-alloc=system flag to configure so that the libc version of malloc() is used instead of one of the internal memory allocator implementations.
Fix a few mis-sorts in BUILD_DEPENDS and don't list www/p5-libwww
twice.

Fix a compile error in the bridges code when --enable-debug is
specified.

Fix a sporadic failure in the clear_001 QA test that occasionally
breaks the build.  The nominal mutex hold time is 5 seconds, but
because it is only checked one per second, the actual time is more
likely to be around 6 seconds.  When the before time value is
subtracted from the after time value, the result might be 7 whole
seconds and a large negative number of nanoseconds.  Since the
pass/fail check only looks at the difference in the seconds fields,
it will fail the "< 7" assertion.  Relax the assertion to "<= 7"
as a quick fix.

The editors/openoffice-4 port does not need poppler.  It is only
needed for the PDF import extension, which is not built in OpenOffice
4.1.x.

Cherrypick some cleanups from PR 216245 [1]:
 * The --without-stlport configure flag is not needed.

 * Pass the -with-build-version flag to configure (but use ${PKGNAME}
   for completeness instead of just  ${PORTNAME}-${PORTVERSION})

 * Tweak a comment in the Makefile.

PR:		216245
Submitted by:	pfg [1]