gnatdroid: Prepare for new cross-compiler
The gnatdroid binutils and compiler master ports have been expanded to
support an upcoming FreeBSD/DF -to- FreeBSD-ARM64 cross-compiler.  This
backports the gcc7 aarch64-*-freebsd* support to gcc6-aux along with a
customization of mine for the gnat compiler.

This also adds the missing LINK_SPEC replacement and removes a
now-obsolete post-extract directory creation.  Apparently the fortran
option has been broken since the rebasing on gcc6-aux, now fixed, but
nobody reported it.