Add gnatcross-aarch64, a slave port of gnatdroid (C/Ada cross-compiler)
This is a FreeBSD/DF cross-compiler targeting FreeBSD/ARM64 (aarch64).
It differs from devel/aarch64-none-elf-gcc in the following ways:
  1) uses FreeBSD/ARM64 headers and libs (true sysroot)
  2) Supports Ada front end
  3) Supports Fortran front end (optionally, default off)
  4) Supports ObjC front end (optionally, default off)

It passes all but 5 ACATS tests:
  1) c34005o (GNAT ICE, expand_shift_1 at expmed.c:2490)
  2) c34007i (GNAT ICE, expand_shift_1 at expmed.c:2490)
  3) cb1010a (stack check related)
  4) cb1010c (stack check related)
  5) cb1010d (stack check related)

Ideally this cross-compiler will lead to a fully native Ada compiler for
FreeBSD/ARM64 which leads to lang/gcc6-aux support, enabling the Ada
framework for that platform.