net-p2p/deluge-cli: fix runtime
Drop local patch, that breaks runtime with libtorrent-rasterbar-python >= 1.1.2.
In libtorrent 1.1.2 high_performance_seed() method returns dict, not object, so
this breaks things. If I understand correctly, in newer version this dict
supposed to be applied with apply_settings() method, that is missing in python
binding. Not bumping PORTREVISION, because of subsequent update is comming.