Add devel/ccache-static and devel/ccache-memcached-static ports.
These are to allow packages to be created for use where a static
ccache is needed to avoid build dependencies.  Poudriere is
growing support for this use case.  While a ccache-static is
not likely as useful, the ccache-memcached-static port will
allow Poudriere to use ccache+memcached in its jails without
needing to bootstrap ccache-memcached and dealing with the
cyclic dependency problem of building that or compromising
with not building it with a compiler cache.