Unbreak after r436971.
Import some commits from kde-workspace's git repository, and add another one to
make sure we are linking against the full path to libXss.so and not just
passing "-lXss" to the linker.

rezny@ has a patch to update kde-workspace to its latest (and last) release in
Phabricator, but I'm landing this one to unbreak the build in the meantime.