Update to new upstream release 0.69.
Upstream change "You can now explicitly configure SSH terminal mode
settings not to be sent to the server, if your server objects to them."

Enable pageant, was missing from the 0.68 upgrade.

Switch to GTK3, no longer compiles with GTK2.

The security fixes contained in 0.69 are only good on Windows, so not
flagging this as a security update.