Switch the dependency on archivers/freeze to default to off.
The dependency on archivers/freeze make it so that amavisd-new can check
inside freeze archives, however there are issues with that port that keeps
it from being built, and that makes amavisd-new not build either.  It is
still possible to install freeze separately and have amavisd-new use it.

Approved by:	flo (maintainer)