Update Amazon SSM Agent to version 2.0.790.0.
This update does not appear to bring any significant functional changes
affecting FreeBSD; but some patches from the ports tree have been merged
upstream so this commit reduces the set of local patches.

This version of the Amazon SSM Agent builds a "ssm-cli" utility.  Since I
have been unable to determine what this utility does, and it seems to be
entirely undocumented, it is not being installed.