Python implementations of the Porter, Porter2, Paice-Husk, and Lovins stemming algorithms for English. These implementations are straightforward and efficient, unlike some Python versions of the same algorithms available on the Web. This package is an extraction of the stemming code included in the Whoosh search engine.
Note that these are *pure Python* implementations. Python wrappers for, e.g.
the Snoball stemmers and the C implementation of the Porter stemmer are
available on PyPI and will be faster if using compiled code is an option for
you.

Stemming algorithms attempt to automatically remove suffixes (and in some
cases prefixes) in order to find the "root word" or stem of a given word. This
is useful in various natural language processing scenarios, such as search.

In general ``porter2`` is the best overall stemming algorithm, but not
necessarily the fastest or most aggressive.

WWW: https://pypi.python.org/pypi/stemming