Add rubygem-ddplugin 1.0.1
ddplugin is a library for managing plugins.

Designing a library so that third parties can easily extend it greatly improves
its usefulness. ddplugin helps solve this problem using plugins, which are
classes of a certain type and with a given identifier (Ruby symbol).

This code was extracted from Nanoc, where it has been in production for years.

WWW: https://github.com/ddfreyne/ddplugin