Upgrade cloudabi and cloudabi-utils to the latest upstream versions.
This release of the CloudABI definitions contains some cleanups, such as
the removal of socket addresses. Though structures like sockaddr are
still defined by the C library, this information is no longer exposed by
runtimes. The reasoning being that additional services for CloudABI,
such as Flower, are going to facilitate this in the future:

https://github.com/NuxiNL/flower

Reviewed by:	bapt
Differential Revision:	https://reviews.freebsd.org/D11734