MFH: r447129 r447530
Add a code block for the qemu-user enabled cross build environment.  When using
this environment in poudriere, CC is not set to the default of /usr/bin/cc and
a cross-compile toolchain is used.  We need to hand edit this so that the run
time configuration for python matches what the FreeBSD base system provides.

PR:		208282
Submitted by:	manu
Approved by:	portmgr (mat)

Pointyhat to me.

A stray '.' somehow made it past my testing.

Do *not* bump portrevision as this only affects the packaging/stage
of these ports on mips/armv6 or other cross compiled targets.

PR:		221202
Reported by:	antoine

Approved by:	ports-secteam (zi)