New port: www/qt5-webengine.
Qt WebEngine provides functionality for rendering regions of dynamic web
content. In other words, this is Chromium's Blink web engine and parts of
Chromium's content layer wrapped by Qt. Qt WebKit is deprecated upstream, and
WebEngine is the future.

This port has been in the works for a very long time, with contributions by
myself, tcberner, Adriaan de Groot and Marie Loise Nolden.

Since most of the port is just a large part of Chromium's code base, this means
sharing some of the pain www/chromium goes through by having to patch a ton of
files. We cannot share the exact same patches because this version of Qt
WebEngine is based on an older Chromium release.

I have verified that the port works, but it is possible that many things are
still wrong or missing, so bug reports are encouraged.

Reviewed by:	Adriaan de Groot <groot@kde.org>, tcberner
Differential Revision:	https://reviews.freebsd.org/D12063