Fix up DEFAULT_URL_HOST/DEFAULT_MAIL_HOST for jailed or external builds.
If the package is built in, for instance, poudriere, or on a computer other
than the one it is later installed on, the DEFAULT_MAIL_HOST and
DEFAULT_URL_HOST variables in the Mailman/Defaults.py file were unsuitable.

Leverage pkg-install to fix these up to be the same as hostname -f at
install time.

[The MFH would require the intermediate update to 2.1.24 to be included.]

MFH:		2017Q3