Make yaml-cpp usable on HEAD for software that wants C++17.
yaml-cpp uses auto_ptr in a small number of places. This is problematic,
as libc++ 5.0.0 (used on HEAD) has auto_ptr disabled when building for
C++17. This causes the build for cloudabi-utils to fail on HEAD.

Instead of removing all of the use of auto_ptr from yaml-cpp, this
commit only changes public header files to make use of unique_ptr.

As I also ran into this issue for CloudABI, here's a bug report I filed
upstream a couple of weeks ago:

https://github.com/jbeder/yaml-cpp/issues/523

Reported by:	pkg-fallout :-C
Reviewed by:	madpilot, mat
Differential Revision:	https://reviews.freebsd.org/D12285