- Update to 4.2.3 - Fix various portlint warnings - Add missing LICENSE_FILE - Update pkg-descr to match recent upstream development
The update was postponed a few times because early 4.2 releases
had some serious regressions on FreeBSD. I know that there are
still a few issues in the current release especially around IPTV
and SAT>IP but they do not seem to be FreeBSD specific. If you
run into regressions please let me know then we can still decide
if we need to keep tvheadend 4.0 around.

Thanks to all people for their patience and the people which helped
during the testing phase especially Rainer and Heiko!

Thanks to:	Rainer Hurling <rhurlin@gwdg.de>,
		Heiko Kirschke <heiko.kirschke@posteo.de>