Enable the DEBUG, GCRYPT and PING options by default.
DEBUG is required for any sort of troubleshooting, GCRYPT is required to
encrypt communication with the server, and PING is one of the most basic
sensors which everybody uses.  My hope is that with these three options,
more people will be able to use collectd without having to recompile it.

Approved by:	maintainer