Patch puppet.rb for locales issue
We are currently installing locales to a directory that is unexpected
by puppet and throws loads of warnings in the logs. Here we patch the location
that looks for the locales location to use the location that we install the
locales to. This has been corrected in newer versions of puppet, and the
sysutils/puppet5 port does not install to this same location.

Approved-by:	mat(co-mentor)