Woraround crash when LC_* is not set
Puppet and MCollective define LC_ALL=C.UTF-8 to avoid this crash, but it's
still a problem for many use cases (e.g. cron job storing facts for
MCollective).

While this does not fix the root cause of the crash that is still under
investigation, this makes facter usable regardless of the user's envrionment.

PR:		222125
Submitted by:	ryanb@honeycomb.net