Update to the 20171119 snapshot of GCC 8.
Remove the setting of CXXFLAGS / STAGE1_CXXFLAGS to -fbracket-depth=512
for armv6 which should not be required here (and is not for armv7 and
aarch64 according to tests done by andreast@).

Use .elif instead of distinct .if sequences for architecture-specific
settings, something I plan on rolling out throughout all lang/gcc* ports.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@454509 35697150-7ecd-e111-bb59-0022644237b5