Add flavor support to portmaster. This version has been lightly tested and supports upgrades from non-flavored port versions based on the information in the MOVED file.
For initial installations of flavored ports, the flavor must be specified
as part of the port origin, e.g. "devel/py-py@py36" for the Python-3.6
version of that port. Dependent ports will automatically be installed with
the correct flavor passed via the dependency mechanism.

It is planned to add a --flavor option to ease flavor selection for ports
that are initially installed with portmaster.

Approved by:	antoine (implicit)