Improve handling of the system architecture in `lang/ocaml'.
OCaml is using ARCH variable throughout its makefiles, which clashes with
our own variable of the same name.  To workaround it, OCAML_ARCH variable
was introduced which is set based on ARCH, and makefiles later patched in
the `post-configure' target.

This approach does not scale well with porting OCaml to new architectures
as being fragile and sometimes making the build non-idempotent under some
circumstances.

Instead of "fixing" makefiles post-configure, rename ARCH to SYSTEM_ARCH:
this avoids the clash altogether and lets configure script to set correct
values for supported architectures in the first place with no adjustments
required afterwards.

Tested on:	i386, amd64, powerpc