Actualize BROKEN_sparc64 statement.  It was (re)marked broken last time in r138465 by mi@ (first time in r109062 by kris@), that's 12.5 years ago, and never revisited ever since (except minor facelifting in r387855 by marino@). The world was largely different back then, with GCC 3.4.4 and OCaml 3.08.3.
Currently it does not build on 10.4-STABLE running under QEMU 2.9.0.

While here, remove stale comment about hardcoded GCC for depend target that
was added in r292093 by bf@ and later *not* removed in r385012 by marino@
together with in-place editing commands it was attributed to.  I messed up
the history with r456400, thinking it still applies yet not performing the
proper archaeology check, and rightfully deserve a pointy hat for that.


git-svn-id: svn+ssh://svn.freebsd.org/ports/head@456574 35697150-7ecd-e111-bb59-0022644237b5